requirements.txt
# =============================================================================
# requirements.txt — Vidya AI Teaching Bot
# Install: py -3.11 -m pip install -r requirements.txt
# =============================================================================

# Core voice pipeline
pipecat-ai[sarvam,openai,webrtc]
pipecat-ai-small-webrtc-prebuilt

# Backend server
fastapi
uvicorn[standard]
httpx

# LLM — free alternative to OpenAI
groq

# Database — user profiles and progress
aiosqlite

# Utilities
python-dotenv
loguru
websockets
